console: Separate client form in several modules
